move create_logger to recode_id.py
