+++ products/zap-opendata/src/recode_id.py
@@ -1,13 +1,13 @@
 import json
 import logging
-import time
+import logging
+import os
 from typing import List
 import pandas as pd
 import requests
-from .client import Client
 
-from .util import create_logger
 from . import CLIENT_ID, SECRET, TENANT_ID, ZAP_DOMAIN
+from .client import Client
 
 RECODE_ID_FIELDS = [
     "primary_applicant",
@@ -77,6 +77,22 @@ def get_headers():
         secret=SECRET,
     )
     return client.request_header
+
+
+def create_logger(logger_name, file_name) -> logging.Logger:
+    if not os.path.exists(".logs/"):
+        os.makedirs(".logs/")
+    logger = logging.getLogger(logger_name)
+    logger.setLevel(logging.DEBUG)
+
+    formatter = logging.Formatter("%(asctime)s:%(name)s: %(message)s")
+    logger.handlers = []
+    file_handler = logging.FileHandler(f".logs/{file_name}")
+    file_handler.setLevel(logging.DEBUG)
+    file_handler.setFormatter(formatter)
+    if not logger.hasHandlers():
+        logger.addHandler(file_handler)
+    return logger
 
 
 class ReuseTracker:
